Fase 3: verificar MLAgentsBridge contra Unity real y arreglar dtype discreto
urc env launch conectado a la escena Basic.unity de ML-Agents (Unity
6000.0.77f1) confirma observation_spec/action_spec correctos
(shape=(20,) / shape=(1,) discreto). Reveló un bug que el
UnityEnvironment falso de los tests no cubría: action_spec() reportaba
dtype="float32" para acciones discretas en vez de "int32". Corregido y
cubierto con una aserción de test.

+++ src/urc/bridges/mlagents_bridge.py
@@ -81,7 +81,7 @@ class MLAgentsBridge(BridgeAdapter):
     def action_spec(self) -> ActionSpec:
         spec = self._env.behavior_specs[self._only_behavior_name()].action_spec
         if spec.is_discrete():
-            return ActionSpec(shape=(spec.discrete_size,), discrete=True)
+            return ActionSpec(shape=(spec.discrete_size,), dtype="int32", discrete=True)
         return ActionSpec(shape=(spec.continuous_size,), discrete=False)
 
     def close(self) -> None:
